Fix hideSidebarView to position view correctly.
Code initially used (CGRect){CGPointZero, self.view.frame.size} which assumes
that the view should go back to 0,0. This may not be true (especially if the
view does not have a NavigationController).

Change transform to offset it by the reverse width amount. This is the correct
reverse transform that should be applied to restore the view correctly.